Add voltage control

+++ mqtt/sensor.go
@@ -13,11 +13,11 @@ func toModel(e *event.SensorEvent, d *model.DeviceModel) *model.SensorEventModel
 	r.DeviceId = d.ID
 	r.RealTime = now
 	r.DeviceTime = time.Time(e.Time)
-	r.Period = uint(e.Energy.Period)
-	r.Power = uint(e.Energy.Power)
-	r.ApparentPower = uint(e.Energy.ApparentPower)
-	r.ReactivePower = uint(e.Energy.ReactivePower)
-	r.Voltage = uint(e.Energy.Voltage)
+	r.Period = e.Energy.Period
+	r.Power = e.Energy.Power
+	r.ApparentPower = e.Energy.ApparentPower
+	r.ReactivePower = e.Energy.ReactivePower
+	r.Voltage = e.Energy.Voltage
 	r.Current = e.Energy.Current
 
 	return r
